Routing & other necessary changes for the project.
