--- landreegistration/src/App.js
@@ -1,24 +1,0 @@
-import React from "react";
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import SidebarMenu from "./datasheet"; 
-
-
-function App() {
-  return (
-    <Router>
-      <div>
-        <Switch>
-          <Route exact path="/" component={Home} />
-          <Route path="/records" component={SidebarMenu} /> {/* For datasheet */}
-          
-          
-          <Route path="*">
-            <h1>404 - Page Not Found</h1>
-          </Route>
-        </Switch>
-      </div>
-    </Router>
-  );
-}
-
-export default App;
