+++ landregistration/src/components/Home.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Home() {
+  return (
+    <div>
+      <h1>Welcome to the Land Registry System</h1>
+    </div>
+  );
+}
+
+export default Home;
