+++ landregistration/src/components/ui/Button.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Button = ({ children, onClick }) => {
+  return <button onClick={onClick}>{children}</button>;
+};
+
+export default Button;
