+++ landregistration/src/components/ui/Card.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Card = ({ children }) => {
+  return <div className="card">{children}</div>;
+};
+
+export default Card;
