+++ landregistration/src/components/ui/Input.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Input = ({ value, onChange }) => {
+  return <input value={value} onChange={onChange} />;
+};
+
+export default Input;
