+++ landregistration/src/components/ui/Table.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Table = ({ children }) => {
+  return <table>{children}</table>;
+};
+
+export default Table;
