+++ package.json
@@ -1,0 +1,9 @@
+{
+  "dependencies": {
+    "@fortawesome/free-solid-svg-icons": "^6.7.2",
+    "@fortawesome/react-fontawesome": "^0.2.2",
+    "jquery": "^3.7.1",
+    "lucide-react": "^0.469.0",
+    "web3": "^4.16.0"
+  }
+}
